exception

+++ src/main/java/com/bicoccaprojects/beerconnect/service/LimitedEditionService.java
@@ -1,6 +1,8 @@
 package com.bicoccaprojects.beerconnect.service;
 
 import com.bicoccaprojects.beerconnect.entity.LimitedEdition;
+import com.bicoccaprojects.beerconnect.exception.beer.BeerNotFoundException;
+import com.bicoccaprojects.beerconnect.exception.beer.NoBeersFoundException;
 import com.bicoccaprojects.beerconnect.repository.LimitedEditionRepository;
 import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -8,42 +10,90 @@ import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.Optional;
+import java.util.SplittableRandom;
 
 @Service
 public class LimitedEditionService extends BeerService {
     @Autowired
     private LimitedEditionRepository limitedEditionRepository;
 
-    @Transactional
-    public Iterable<LimitedEdition> getLEBeers() {
-        return  limitedEditionRepository.findAll();
+    public Iterable<LimitedEdition> getAllLEBeers() throws Exception {
+        Iterable<LimitedEdition> beers = limitedEditionRepository.findAll();
+        if (!beers.iterator().hasNext()) {
+            throw new Exception("There are no Limited edition beers in the DB");
+        }
+        return beers;
     }
 
-    @Transactional
-    public Optional<LimitedEdition> getLEBeer(Long id) {
-        return limitedEditionRepository.findById(id);
+    public LimitedEdition getLEBeer(Long id) {
+        Optional<LimitedEdition> optionalBeer = limitedEditionRepository.findById(id);
+        if (optionalBeer.isPresent()){
+            return optionalBeer.get();
+        } else {
+            throw new BeerNotFoundException(id);
+        }
     }
 
     @Transactional
     public void deleteLEBeer(Long id) {
+        Optional<LimitedEdition> beerOptional = limitedEditionRepository.findById(id);
+        if (beerOptional.isEmpty()) {
+            throw new BeerNotFoundException(id);
+        }
         limitedEditionRepository.deleteById(id);
     }
 
     @Transactional
     public void deleteLEBeers() {
+        Iterable<LimitedEdition> beers = limitedEditionRepository.findAll();
+        if (!beers.iterator().hasNext()) {
+            throw new NoBeersFoundException();
+        }
         limitedEditionRepository.deleteAll();
     }
 
     @Transactional
-    public void addLEBeer(LimitedEdition leBeer) {
-        limitedEditionRepository.save(leBeer);
+    public void addLEBeer(LimitedEdition leBeer) throws Exception {
+        Long id = leBeer.getIdBeer();
+        Optional<LimitedEdition> beerOptional = limitedEditionRepository.findById(id);
+        if(beerOptional.isEmpty()){
+            List<String> beerList = limitedEditionRepository.searchBeerByName(leBeer.getNameBeer());
+            if(beerList.isEmpty()){
+                limitedEditionRepository.save(leBeer);
+            }else {
+                throw new Exception("There is already a beer with "+leBeer.getNameBeer()+" as name");
+            }
+        }else {
+            throw new Exception("There is already a beer in the DB with ID "+(leBeer.getIdBeer()).toString());
+        }
     }
 
     @Transactional
     public void updateLEBeer(LimitedEdition leBeer) {
-        limitedEditionRepository.save(leBeer);
+        Long id = leBeer.getIdBeer();
+        Optional<LimitedEdition> beerOptional = limitedEditionRepository.findById(id);
+        if(beerOptional.isPresent()){
+            limitedEditionRepository.save(leBeer);
+        }else {
+            throw new BeerNotFoundException(id);
+        }
     }
 
-    @Transactional
-    public List<String> findLEBeer(String originalBeer){ return limitedEditionRepository.findLEByBeer(originalBeer); }
+    public List<String> findLEByOriginalBeer(String originalBeer){
+        List<String> beerList = limitedEditionRepository.findLEByBeer(originalBeer);
+        if (!beerList.isEmpty()){
+            return beerList;
+        }else{
+            throw new BeerNotFoundException("There are no beer with this name");
+        }
+    }
+
+    public List<String> searchLEByName(String nameBeer){
+        List<String> beerList = limitedEditionRepository.searchBeerByName(nameBeer);
+        if(!beerList.isEmpty()){
+            return beerList;
+        }else {
+            throw new BeerNotFoundException("There are no beer with this name");
+        }
+    }
 }
